Introduce a new 'clockSynced' animation component property
when set to true (default is false), the animation will be syncronized
with the browsers's clock. The effect of this is that, assuming all
viewers clocks are correct, everybody will see the animation in the
same state, as if it was an event happening at the same time for all
the viewers.